requirements.txt
python-dotenv
gpiozero
git+https://github.com/nnewman/pysnoo@develop#egg=pysnoo
